Use Vue proxy when updating Importer
Because the JSZip is async, updates to self.status happen after we've returned.  This means that Vue has seen the "Ready for Import" we set before loading the zip, and not the updates that might show errors.

This makes async writes go through the Vue Proxy, and therefore trigger refreshes.